pyproject.toml
[tool.poetry]
name = "fastapi-agent"
version = "0.1.0-beta"
description = "An intelligent terminal agent with LLM integration, expert modes, and command execution"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "0.104.1"
uvicorn = "0.24.0"
httpx = "0.25.2"
rich = "13.7.0"
pydantic = "2.5.3"
pydantic-settings = "2.1.0"
pyyaml = "^6.0"
prompt-toolkit = "^3.0.43"

[tool.poetry.scripts]
agent-cli = "src.cli.interactive_terminal:main"
agent-api = "src.api.main:app"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
